modal de dele usuarios

+++ Tera_Web/Tera_Web/Controllers/UserController.cs
@@ -127,15 +127,20 @@ namespace Tera_Web.Controllers
         }
 
         // GET: UsuarioController/Delete/5
-        public ActionResult Delete(int id)
+        [HttpPost]
+        public ActionResult Delete(int id, bool confirmed)
         {
-            userModel.DeleteUsers(id);
+            if (confirmed)
+            {
+                userModel.DeleteUsers(id);
+            }
+
             return RedirectToAction("List", "User");
         }
 
         // POST: UsuarioController/Delete/5
         [HttpPost]
-        [ValidateAntiForgeryToken]
+
         public ActionResult Delete(int id, IFormCollection collection)
         {
             try
